requirements-gmail.txt
google-auth>=2.28
google-auth-oauthlib>=1.2
google-api-python-client>=2.120
